content-type: Move the ContentType Type in its own module
This add more clarity to the code base
